+++ src/classes/Customer.java
@@ -23,7 +23,7 @@ public class Customer implements Serializable{
     private String city;
     private String state;
     private int zip;
-    private int phone;
+    private long phone;
     private String email;
     
     //CONSTRUCTORES
@@ -148,14 +148,14 @@ public class Customer implements Serializable{
     /**
      * @return the phone
      */
-    public int getPhone() {
+    public long getPhone() {
         return phone;
     }
 
     /**
      * @param phone the phone to set
      */
-    public void setPhone(int phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
